Find external property assertions for individuals from ontology.
The xslt parser cannot extract metadata from other ontologies. This
adds a parser to process the external properties and if a match
is found then the appropiate property is added (dp,op,ni). If no
match is found then add in the documentation an "ep" property
which indicates that the property is found in an external
ontology.